moved all logic that was temproarly in one file into separate files on the server
