+++ apps/server/src/auth/utils/pass-utils.ts
@@ -1,19 +1,10 @@
 // Not to confuse with jsonwebtoken verify
-import { hash, verify as verifyArgon2 } from "argon2";
+import {hash, verify as verifyArgon2} from "argon2";
 
 export class PassUtils {
-	// eslint-disable-next-line no-use-before-define
-	private static _instance: PassUtils;
-
-	private constructor() {}
-
-	static get instance() {
-		if (!this._instance) {
-			this._instance = new PassUtils();
-		}
-
-		return this._instance;
+	private constructor() {
 	}
+
 
 	static async hash(unhashed: string) {
 		return hash(unhashed);
