+++ apps/server/src/auth/validation/email-already-in-use.ts
