+++ apps/server/src/middleware/error-handler.ts
@@ -1,15 +1,15 @@
-import { CustomError } from "../errors/custom-error";
-import { SomethingWentWrongResponse } from "../responses/internal-server-error";
-import { ErrorHandler } from "../types";
-import { logger } from "../utils/logger";
+import {CustomError} from "../errors/custom-error";
+import {SomethingWentWrongResponse} from "../responses/internal-server-error";
+import {ErrorHandler} from "../types";
+import {logger} from "../utils/logger";
 
-const errorHandler: ErrorHandler = (err, _req, res, _next) => {
+export const errorHandler: ErrorHandler = (err, _req, res, _next) => {
 	if (err instanceof CustomError) {
 		const errors = err.serializeErrors();
 
 		logger.error(errors);
 
-		return res.status(err.statusCode).send({ errors, data: null });
+		return res.status(err.statusCode).send({errors, data: null});
 	}
 
 	logger.error(err);
@@ -19,4 +19,3 @@ const errorHandler: ErrorHandler = (err, _req, res, _next) => {
 	});
 };
 
-export { errorHandler };
