+++ apps/server/src/middleware/restful/types.ts
@@ -1,0 +1,4 @@
+import {Method} from "@prioritea/types";
+import {RequestHandler} from "express";
+
+export type Restful = (methods: Array<Method>) => RequestHandler;
