+++ apps/server/src/tasks/tasks.service.ts
@@ -1,36 +1,8 @@
 import {CreateTaskDto, UpdateTaskDto} from "@prioritea/types";
-import {Task} from "@prisma/client";
 import {Service} from "../core/service";
 import {TasksRepository} from "./tasks.repository";
-import {PrismaClientKnownRequestError} from "@prisma/client/runtime";
-import {NotFoundError} from "../errors/not-found-error";
-
-export interface ITasksService {
-	createTask(
-		userId: string,
-		task: CreateTaskDto
-	): Promise<Array<Task> | null>;
-	getTasks(userId: string): Promise<Array<Task> | null>;
-	getTask(id: string): Promise<Task | null>;
-	updateTask(
-		userId: string,
-		task: UpdateTaskDto
-	): Promise<Array<Task> | null>;
-	deleteTasks(userId: string, ids: Array<string>): Promise<{
-		count: number;
-		tasks:Array<Task> | null;
-	}>;
-}
-
-export const updatedNonExistentTask = function (err: unknown) {
-	if (
-		err instanceof PrismaClientKnownRequestError &&
-		err.code === "P2025" &&
-		err.meta?.cause === "Record to update not found."
-	) {
-		throw new NotFoundError("No task matches the provided id.");
-	}
-};
+import {ITasksService} from "./interfaces";
+import {updatedNonExistentTask} from "./validation/updated-non-existent-task";
 
 
 export class TasksService extends Service<TasksRepository> implements ITasksService {
