+++ apps/server/src/tasks/utils/id-param-fallback.ts
@@ -1,0 +1,5 @@
+import {IdParamError} from "../../errors/id-param-error";
+
+export const idParamFallback = async () => {
+	throw new IdParamError();
+};
