+++ apps/server/src/tasks/validation/id-schema.ts
@@ -1,0 +1,5 @@
+import {z} from "zod";
+
+export const idSchema = z.object({
+	id: z.string().cuid(`id must be a valid CUID`),
+});
