+++ apps/server/src/tests/.env.test
@@ -1,1 +1,1 @@
-"postgresql://postgres:BFDH7F8532A13B48A47869AA748F9E6EC04F4E727F8AB0FE47E6D0E9CF263F4A@localhost:5432/test-prioritea?schema=public"
+DB_URL="//postgres:BFDH7F8532A13B48A47869AA748F9E6EC04F4E727F8AB0FE47E6D0E9CF263F4A@localhost:5432/test-prioritea?schema=public"
