+++ apps/server/src/tests/fixtures.ts
@@ -3,6 +3,7 @@ import {VITE_PORT} from "../env/env";
 import supertest from "supertest";
 
 const app = new App(Number(VITE_PORT)).init().app;
+
 export const request = supertest(app);
